# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Finwiz",
        "description": "Finwiz API documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/user/user-assets/{accountId}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get user's Assets",
                "operationId": "68742f35d30742cbee1c4ef57b3008c4",
                "parameters": [
                    {
                        "name": "accountId",
                        "in": "path",
                        "description": "ID of the user user'Account to retrieve assets",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/user-assets/{id}/{account_id}": {
            "get": {
                "tags": [
                    "Assets"
                ],
                "summary": "Get a user's Specific asset By ID",
                "operationId": "2cc801a78eb9ed48f56aebc1750405be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user Asset to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "Account Id of the user Asset to retrieve",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/user-assets": {
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Create a user assets",
                "operationId": "7aa5110813f02c42c91f311aa59189a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "account_id": {
                                        "type": "integer",
                                        "example": "12"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": "20"
                                    },
                                    "label": {
                                        "type": "varchar",
                                        "example": "Account Holder"
                                    },
                                    "avatar": {
                                        "description": "Image file to upload",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/user-assets/{id}": {
            "post": {
                "tags": [
                    "Assets"
                ],
                "summary": "Update a user's assets by ID",
                "operationId": "4c7bdd4452372d5c671dad24f2eac73f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user Asset to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": "20"
                                    },
                                    "label": {
                                        "type": "varchar",
                                        "example": "Account Holder"
                                    },
                                    "avatar": {
                                        "description": "Image file to upload",
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Assets"
                ],
                "summary": "Delete a user Assets by ID",
                "operationId": "cb27f072fd62ebf52bde4a5b4d0de79b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user's Assets to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID of the user user'Account to retrieve assets",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/user/categories/{per_page}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "operationId": "19efdfd902a7d133add9cc20c286def8",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        },
        "/api/user/main-categories/{per_page}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all main categories",
                "operationId": "c6048e42dd820dacec033ffb02e79619",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        },
        "/api/user/sub-categories/{parent_id}/{per_page}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get sub-categories by parent ID",
                "operationId": "17eed4e71c074ab425f5f55429022506",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "path",
                        "description": "ID of the parent category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "path",
                        "description": "Number of items per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "default": 25
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success response"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization ",
                "in": "header",
                "scheme": "bearer"
            }
        }
    ]
}